requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
anyio==3.7.1
asks==3.0.0
async-generator==1.10
asyncio==4.0.0
attrs==25.4.0
certifi==2025.11.12
curio-compat==1.6.7
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
multidict==6.7.0
outcome==1.3.0.post0
propcache==0.4.1
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.32.0
typing_extensions==4.15.0
yarl==1.22.0
